# backend/openapi.json
{"openapi":"3.1.0","info":{"title":"backend","description":"","license":{"name":""},"version":"0.1.0"},"paths":{"/v1/player_stats":{"get":{"tags":[],"operationId":"get_player_stats_list","responses":{"200":{"description":"玩家统计","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/PlayerStats"}}}}}}}},"/v1/player_stats/{player_id}":{"get":{"tags":[],"operationId":"get_player_stats","parameters":[{"name":"player_id","in":"path","description":"玩家ID","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"玩家统计和比赛详情","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PlayerStatsWithGames"}}}},"404":{"description":"玩家未找到"}}}}},"components":{"schemas":{"GameDetail":{"type":"object","required":["game_id","tournament_name","tournament_sub_name","tournament_date","tournament_location","session_name","players"],"properties":{"game_id":{"type":"integer","format":"int32"},"players":{"type":"array","items":{"$ref":"#/components/schemas/PlayerGameResult"}},"session_name":{"type":"string"},"tournament_date":{"type":"string","format":"date"},"tournament_location":{"type":"string"},"tournament_name":{"type":"string"},"tournament_sub_name":{"type":"string"}}},"PlayerGameResult":{"type":"object","required":["player_id","player_name","game_point","table_point","place_point"],"properties":{"game_point":{"type":"integer","format":"int32"},"place_point":{"type":"integer","format":"int32"},"player_id":{"type":"integer","format":"int32"},"player_name":{"type":"string"},"table_point":{"type":"integer","format":"int32"}}},"PlayerStats":{"type":"object","required":["player_id","player_name"],"properties":{"avg_gp":{"type":["number","null"],"format":"float"},"avg_rp":{"type":["number","null"],"format":"float"},"avg_tp":{"type":["number","null"],"format":"float"},"first_place_count":{"type":["integer","null"],"format":"int64"},"first_place_ratio":{"type":["number","null"],"format":"float"},"fourth_place_count":{"type":["integer","null"],"format":"int64"},"fourth_place_ratio":{"type":["number","null"],"format":"float"},"game_count":{"type":["integer","null"],"format":"int64"},"player_id":{"type":"integer","format":"int32"},"player_name":{"type":"string"},"second_place_count":{"type":["integer","null"],"format":"int64"},"second_place_ratio":{"type":["number","null"],"format":"float"},"third_place_count":{"type":["integer","null"],"format":"int64"},"third_place_ratio":{"type":["number","null"],"format":"float"},"total_gp":{"type":["integer","null"],"format":"int64"},"total_rp":{"type":["integer","null"],"format":"int64"},"total_tp":{"type":["integer","null"],"format":"int64"}}},"PlayerStatsWithGames":{"type":"object","required":["player_id","player_name","game_details"],"properties":{"avg_gp":{"type":["number","null"],"format":"float"},"avg_rp":{"type":["number","null"],"format":"float"},"avg_tp":{"type":["number","null"],"format":"float"},"first_place_count":{"type":["integer","null"],"format":"int64"},"first_place_ratio":{"type":["number","null"],"format":"float"},"fourth_place_count":{"type":["integer","null"],"format":"int64"},"fourth_place_ratio":{"type":["number","null"],"format":"float"},"game_count":{"type":["integer","null"],"format":"int64"},"game_details":{"type":"array","items":{"$ref":"#/components/schemas/GameDetail"}},"player_id":{"type":"integer","format":"int32"},"player_name":{"type":"string"},"second_place_count":{"type":["integer","null"],"format":"int64"},"second_place_ratio":{"type":["number","null"],"format":"float"},"third_place_count":{"type":["integer","null"],"format":"int64"},"third_place_ratio":{"type":["number","null"],"format":"float"},"total_gp":{"type":["integer","null"],"format":"int64"},"total_rp":{"type":["integer","null"],"format":"int64"},"total_tp":{"type":["integer","null"],"format":"int64"}}}}}}